added dynamic routes with names

+++ routes/web.php
@@ -15,36 +15,36 @@ use Illuminate\Support\Facades\Route;
 
 Auth::routes();
 
-Route::get('/admin', 'UsersController@index')->middleware('role:admin');
-Route::get('/admin/{user}/edit', 'UsersController@edit')->middleware('role:admin');
+Route::get('/admin', 'UsersController@index')->name('admin.index')->middleware('role:admin');
+Route::get('/admin/{user}/edit', 'UsersController@edit')->name('admin.edit')->middleware('role:admin');
 Route::patch('/admin/{user}', 'UsersController@update')->name('admin.update')->middleware('role:admin');
-Route::delete('/admin/{user}', 'UsersController@destroy')->middleware('role:admin');
+Route::delete('/admin/{user}', 'UsersController@destroy')->name('admin.destroy')->middleware('role:admin');
 
-Route::get('/courses/create', 'CoursesController@create')->middleware('role:admin,instructor');
-Route::post('/courses/store', 'CoursesController@store')->middleware('role:admin,instructor');
+Route::get('/courses/create', 'CoursesController@create')->name('courses.create')->middleware('role:admin,instructor');
+Route::post('/courses/store', 'CoursesController@store')->name('courses.store')->middleware('role:admin,instructor');
 Route::get('/courses/{course}/edit', 'CoursesController@edit')->name('courses.edit')->middleware('role:admin,instructor');
 Route::patch('/courses/{course}', 'CoursesController@update')->name('courses.update')->middleware('role:admin,instructor');
-Route::get('/courses/{course}', 'CoursesController@show')->name('courses.show');
-Route::delete('/courses/{course}', 'CoursesController@destroy')->middleware('role:admin,instructor');
+Route::get('/courses/{course}', 'CoursesController@index')->name('courses.index');
+Route::delete('/courses/{course}', 'CoursesController@destroy')->name('courses.destroy')->middleware('role:admin,instructor');
 
-Route::get('/lessons/create/{course}', 'LessonsController@create')->middleware('role:admin,instructor');
-Route::post('/lessons/store', 'LessonsController@store')->middleware('role:admin,instructor');
+Route::get('/lessons/create/{course}', 'LessonsController@create')->name('lessons.create')->middleware('role:admin,instructor');
+Route::post('/lessons/store', 'LessonsController@store')->name('lessons.store')->middleware('role:admin,instructor');
 Route::get('/lessons/{lesson}/edit', 'LessonsController@edit')->name('lessons.edit')->middleware('role:admin,instructor');
 Route::patch('/lessons/{lesson}', 'LessonsController@update')->name('lessons.update')->middleware('role:admin,instructor');
-Route::get('/lessons/{lesson}', 'LessonsController@show')->name('lessons.show');
-Route::delete('/lessons/{lesson}', 'LessonsController@destroy')->middleware('role:admin,instructor');
-Route::post('/lessons/image/upload', 'LessonsController@uploadImage');
+Route::get('/lessons/{lesson}', 'LessonsController@index')->name('lessons.index');
+Route::delete('/lessons/{lesson}', 'LessonsController@destroy')->name('lessons.destroy')->middleware('role:admin,instructor');
+Route::post('/lessons/image/upload', 'LessonsController@uploadImage')->name('lessons.upload');
 
-Route::post('/reply/store', 'CommentsController@reply');
-Route::post('/comments/store', 'CommentsController@store');
-Route::delete('/comments/{comment}', 'CommentsController@destroy');
-Route::put('/comments/{comment}', 'CommentsController@approved')->middleware('role:admin');
-Route::get('/comments/{comment}/edit', 'CommentsController@edit');
-Route::patch('/comments/{comment}', 'CommentsController@update');
-Route::patch('/comments/flag/{comment}', 'CommentsController@flag');
-Route::get('/comments', 'CommentsController@index')->middleware('role:admin');
+Route::post('/reply/store', 'CommentsController@reply')->name('comments.reply');
+Route::post('/comments/store', 'CommentsController@store')->name('comments.store');
+Route::delete('/comments/{comment}', 'CommentsController@destroy')->name('comments.destroy');
+Route::put('/comments/{comment}', 'CommentsController@approved')->name('comments.approved')->middleware('role:admin');
+Route::get('/comments/{comment}/edit', 'CommentsController@edit')->name('comments.edit');
+Route::patch('/comments/{comment}', 'CommentsController@update')->name('comments.update');
+Route::patch('/comments/flag/{comment}', 'CommentsController@flag')->name('comments.flag');
+Route::get('/comments', 'CommentsController@index')->name('comments.index')->middleware('role:admin');
 
-Route::get('/profile/{user}', 'ProfilesController@index')->name('profile.show');
+Route::get('/profile/{user}', 'ProfilesController@index')->name('profile.index');
 Route::get('/profile/{user}/edit', 'ProfilesController@edit')->name('profile.edit');
 Route::patch('/profile/{user}', 'ProfilesController@update')->name('profile.update');
